% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correctp}
\alias{correctp}
\title{Parameter validation for SPLS}
\arguments{
\item{X}{Predictor matrix}

\item{Y}{Response matrix}

\item{eta}{Sparsity parameter between 0 and 1}

\item{K}{Number of components}

\item{kappa}{Parameter between 0 and 0.5}

\item{select}{Selection method ("pls2" or "simpls")}

\item{fit}{Fitting method}
}
\value{
List with validated parameters
}
\description{
Validates and corrects parameters for the SPLS algorithm
}
