% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_spls_cpp}
\alias{cv_spls_cpp}
\title{Cross-Validation for SPLS}
\usage{
cv_spls_cpp(
  x,
  y,
  fold,
  eta,
  K,
  kappa,
  select,
  scale_x,
  scale_y,
  eps,
  maxstep,
  trace
)
}
\arguments{
\item{x}{Predictor matrix (n x p)}

\item{y}{Response matrix (n x q)}

\item{fold}{Number of folds for cross-validation}

\item{eta}{Numeric vector of sparsity parameters}

\item{K}{Integer vector of candidate numbers of components}

\item{kappa}{Sparsity parameter}

\item{select}{Selection method}

\item{scale_x}{Logical, whether to scale x}

\item{scale_y}{Logical, whether to scale y}

\item{eps}{Convergence criterion}

\item{maxstep}{Maximum number of iterations}

\item{trace}{Logical, whether to print progress}
}
\value{
A list containing:
\item{mspemat}{Matrix of mean squared prediction errors (MSPE)}
\item{eta.opt}{Optimal eta value}
\item{K.opt}{Optimal number of latent components}
}
\description{
This function performs cross-validation for the number of
latent components and sparsity parameters
}
