\name{GeneSPLS-package}
\alias{GeneSPLS-package}
\alias{GeneSPLS}
\docType{package}
\title{
\packageTitle{GeneSPLS}
}
\description{
\packageDescription{GeneSPLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GeneSPLS}
\packageIndices{GeneSPLS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GeneSPLS}

Maintainer: \packageMaintainer{GeneSPLS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
