% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spls_dv}
\alias{spls_dv}
\title{Sparse PLS Direction Vector Estimation}
\usage{
spls_dv(Z, eta, kappa, eps, maxstep)
}
\arguments{
\item{Z}{Cross-product matrix (p x q)}

\item{eta}{Sparsity parameter between 0 and 1}

\item{kappa}{Parameter between 0 and 0.5}

\item{eps}{Convergence criterion}

\item{maxstep}{Maximum number of iterations}
}
\value{
Direction vector
}
\description{
Calculates the direction vector for Sparse PLS
}
