% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ust}
\alias{ust}
\title{Univariate Soft Thresholding (UST) Function}
\arguments{
\item{b}{Vector to be thresholded}

\item{eta}{Threshold parameter between 0 and 1}
}
\value{
Soft thresholded vector
}
\description{
Performs soft thresholding on a vector
}
