% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{widekernelpls_fit}
\alias{widekernelpls_fit}
\title{Wide Kernel PLS Algorithm Implementation}
\usage{
widekernelpls_fit(X, Y, ncomp, center = TRUE, tol = 0, maxit = 100L)
}
\arguments{
\item{X}{The predictor matrix}

\item{Y}{The response matrix}

\item{ncomp}{Number of components to extract}

\item{center}{Logical, whether to center X and Y (default: true)}

\item{tol}{Numeric, tolerance used for determining convergence (default: sqrt(machine_eps))}

\item{maxit}{Integer, maximum number of iterations (default: 100)}
}
\value{
A list containing only:
\item{coefficients}{The regression coefficients matrix}
\item{projection}{The projection matrix}
\item{Xmeans}{The column means of X (if centered)}
\item{Ymeans}{The column means of Y (if centered)}
}
\description{
This function implements the wide kernel PLS algorithm as described by
Rannar et al. (1994). It's optimized for datasets with more variables than observations.
}
