% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_spls_cpp}
\alias{cv_spls_cpp}
\title{Cross-Validation for SPLS}
\usage{
cv_spls_cpp(
  x,
  y,
  fold,
  eta,
  K,
  kappa,
  select,
  scale_x,
  scale_y,
  eps,
  maxstep,
  trace
)
}
\arguments{
\item{x}{Predictor matrix (n x p)}

\item{y}{Response matrix (n x q)}

\item{fold}{Number of folds for cross-validation (default: 5)}

\item{eta}{Numeric vector of sparsity parameters (default: (0.95, 0.96, 0.97, 0.98, 0.99))}

\item{K}{Integer vector of candidate numbers of components (default: (1, 2, 3, 4, 5))}

\item{kappa}{Sparsity parameter (default: 0.5)}

\item{select}{Selection method (default: "pls2")}

\item{scale_x}{Logical, whether to scale x (default: TRUE)}

\item{scale_y}{Logical, whether to scale y (default: FALSE)}

\item{eps}{Convergence criterion (default: 1e-4)}

\item{maxstep}{Maximum number of iterations (default: 100)}

\item{trace}{Logical, whether to print progress (default: FALSE)}
}
\value{
A list containing:
\item{mspemat}{Matrix of mean squared prediction errors (MSPE)}
\item{eta.opt}{Optimal eta value}
\item{K.opt}{Optimal number of latent components}
}
\description{
This function performs cross-validation for the number of
latent components and sparsity parameters
}
