% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spls_cpp}
\alias{spls_cpp}
\title{Sparse Partial Least Squares Regression}
\usage{
spls_cpp(
  x,
  y,
  K,
  eta,
  kappa = 0.5,
  select = "pls2",
  fit = "widekernelpls",
  scale_x = TRUE,
  scale_y = FALSE,
  eps = 1e-04,
  maxstep = 100L,
  trace = FALSE
)
}
\arguments{
\item{x}{Predictor matrix (n x p)}

\item{y}{Response matrix (n x q)}

\item{K}{Number of latent components}

\item{eta}{Sparsity parameter, 0 < eta < 1}

\item{kappa}{If y is multivariate, 0 < kappa <= 0.5}

\item{select}{Method for variable selection: "pls2" (update Y) or "simpls" (update X)}

\item{fit}{Always "widekernelpls" in this implementation}

\item{scale_x}{Whether to scale the predictor matrix}

\item{scale_y}{Whether to scale the response matrix}

\item{eps}{Convergence criterion for direction vector calculation}

\item{maxstep}{Maximum number of iterations for direction vector calculation}

\item{trace}{Whether to print progress information}
}
\value{
A list containing model information including:
\item{betahat}{Regression coefficients}
\item{A}{Active set of predictors}
\item{projection}{Projection matrix}
}
\description{
C++ implementation of the Sparse Partial Least Squares algorithm
}
